restructure project
Combine api, model and client modules with server module. The goal is to reduce build problems with multi-module project, where (for the forseeable future) no other project / code will need one of those modules as a dependency by itself. Rather, when building on Cinnamon, it will be easier just to depend on the whole project.