cleanup tests, wip
Moving from creating test objects via SQL inserts to creating them via the CinnamonClient.

This has the advantage of easier refactoring of tests and having all state of a test case in one place (the @Test in the test class) instead of parts in the sql script and parts in the test class.