changes on indexItems are communicated directly to the IndexService
Background: during testing, a new IndexItem is created - and it was not always picked up by the index thread in time to successfully complete the search test. Now, if an IndexItem is changed (create/update/delete), the servlet responsible updates the IndexService's list of items directly, removing the need to refresh the list of items every couple seconds.