remove lock(osd) requirement for osdUpdate
Before you had to lock an object to change its metadata, now it only checks that the object is not locked by another user.

Admins are exempt from the lock requirement.

refactoring: check the lockedByOtherUser status of all OSDs that the client wants to change before starting on making any changes.

close #425